fixup move tests
